feat: integrate jina-auth
