+++ .github/workflows/cd.yml
@@ -42,15 +42,10 @@ jobs:
           python -m pip install wheel
           pip install --no-cache-dir ".[full,test]"
           sudo apt-get install libsndfile1
-      - name: Add JCloud auth token
-        run: |
-          mkdir -p ~/.jina
-          touch ~/.jina/config.json
-          echo "{\"auth_token\": \"${CI_TOKEN}\"}" > ~/.jina/config.json
-        env:
-          CI_TOKEN: ${{ secrets.CI_TOKEN }}
       - name: Test
         id: test
+        env:
+          JINA_AUTH_TOKEN: ${{ secrets.CI_TOKEN }}
         run: |
           pytest --suppress-no-test-exit-code --cov=jcloud --cov-report=xml \
             -v -s --log-cli-level=DEBUG -m "not gpu" ${{ matrix.test-path }}
