+++ .github/workflows/ci.yml
@@ -132,13 +132,6 @@ jobs:
           export JCLOUD_LOGLEVEL=DEBUG
           pip install --no-cache-dir ".[full,test]"
           sudo apt-get install libsndfile1
-      - name: Add JCloud auth token
-        run: |
-          mkdir -p ~/.jina
-          touch ~/.jina/config.json
-          echo "{\"auth_token\": \"${CI_TOKEN}\"}" > ~/.jina/config.json
-        env:
-          CI_TOKEN: ${{ secrets.CI_TOKEN }}
       - name: Test
         id: test
         run: |
@@ -147,6 +140,7 @@ jobs:
           echo "::set-output name=codecov_flag::jcloud"
         timeout-minutes: 30
         env:
+          JINA_AUTH_TOKEN: ${{ secrets.CI_TOKEN }}
           WOLF_TOKEN: ${{ secrets.WOLF_TOKEN }}
           ARTIFACT_TOKEN: ${{ secrets.ARTIFACT_TOKEN }}
       - name: Check codecov file
