+++ .github/workflows/integration-tests.yml
@@ -27,6 +27,8 @@ jobs:
   integration-tests:
     needs: prep-testbed
     runs-on: ubuntu-latest
+    env:
+      JINA_AUTH_TOKEN: ${{ secrets.CI_TOKEN }}
     strategy:
       fail-fast: false
       matrix:
@@ -46,13 +48,6 @@ jobs:
           python -m pip install wheel
           pip install --no-cache-dir ".[test]"
           sudo apt-get install libsndfile1
-      - name: Add JCloud auth token
-        run: |
-          mkdir -p ~/.jina
-          touch ~/.jina/config.json
-          echo "{\"auth_token\": \"${CI_TOKEN}\"}" > ~/.jina/config.json
-        env:
-          CI_TOKEN: ${{ secrets.CI_TOKEN }}
       - name: Test
         id: test
         run: |
